pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '\.pyi?$'

[tool.mypy]
python_version = 3.11
ignore_missing_imports = true

[tool.poetry]
name = "rfm"
version = "0.0.1"
authors = ["Piotr Gaiński"]
description = "Retro Feasibility Model"

[tool.poetry.dependencies]
torch-geometric = '2.5.3'
torchmetrics = '1.2.0'
torchtyping = '0.1.4'
dgllife = '0.3.2'
tqdm = '4.66.1'
wandb = '0.15.12'
typing-extensions = '4.8.0'
pre-commit = '3.5.0'
gin-config = '0.5.0'
pandas = '2.1.1'
rdkit = '2023.3.3'
pydantic = '1.10.5'
pytest = '7.4.2'
numpy = '1.23.2'
pytorch-lightning = '2.4.0'
